requirements.txt
gradio==4.8.0
langchain==0.0.348
python-dotenv==1.0.0
